+++ __tests__/unit/components/Input/InputFee.spec.js
@@ -340,7 +340,7 @@ describe('InputFee', () => {
 
       describe('when the balance is bigger than the fee', () => {
         it('should return an empty message', () => {
-          // NOTE: Balance is in arktoshi, while fee is in ARK
+          // NOTE: Balance is in arktoshi, while fee is in PHANTOM
           wrapper.vm.wallet_fromRoute.balance = 20e8
           wrapper.vm.fee = 1
           expect(wrapper.vm.insufficientFundsError).toBeNull()