+++ __tests__/unit/pages/Wallet/WalletAll.spec.js
@@ -36,7 +36,7 @@ describe('pages > WalletAll', () => {
           }
         },
         session_network: {
-          symbol: 'Ѧ',
+          symbol: 'ⓟ',
           market: {
             enabled: true
           }