+++ __tests__/unit/store/modules/peer.spec.js
@@ -50,8 +50,8 @@ describe('peer store module', () => {
   })
 
   it('should get random seed server peer', () => {
-    const randomSeedPeers = store.getters['peer/randomSeedPeers'](5, 'ark.mainnet')
-    const randomSeedPeers2 = store.getters['peer/randomSeedPeers'](5, 'ark.mainnet')
+    const randomSeedPeers = store.getters['peer/randomSeedPeers'](5, 'phantom.mainnet')
+    const randomSeedPeers2 = store.getters['peer/randomSeedPeers'](5, 'phantom.mainnet')
     expect(randomSeedPeers).not.toEqual(randomSeedPeers2)
   })